Moved some CommonCheckButton code to common CheckButton
